Fix loading the default open encoding option
This got broken by 907a79263d907fa354d3b4c6570f33047801dd84 back in
2011 as encoding names started to be compared more permissively,
leading to "none" matching the "None" encoding.

This shouldn't be too much of a problem as trying the None encoding
first should not cause any issue, but fix loading of the option anyway.

This however won't restore settings from existing configuration files
that used the previous code, as there is no way to tell whether the
user selected the None encoding voluntarily or not.