Nice on-new-line bracket handling
When using auto closing brackets, the cursor is between the opening and
closing bracket. When hitting enter for new-line, the closing bracket is
still behind the cursor.
If the cursor is between an opening and closing bracket and enter is
hit, in my opinion it, it is desired to get a situation like this (where
| is the cursor)

{
    |
}

or

void main() {
    if( true ) {
        |
    }
}